Add testing files in a separate folder
